Remove dependency on wget
The script does not need to depend on *both* `curl` and `wget`. Not only
is this redundant, but this also allows the script to be run from an EdgeRouter
(i.e., vanilla EdgeOS) without installing any additional packages.

Also, `wget` is arguably less common in OOTB/vanilla installs of Linux distros
as compared to `curl`, which is _nearly_ guaranteed (except for the tiniest distros).